renamed webapp_comms to network_communication
