+++ src/network_communication/CMakeLists.txt
