+++ src/network_communication/scripts/webcam_image_subscriber.py
