+++ src/network_communication/src/archived_code/dummy_image_publisher.cpp
