+++ src/network_communication/src/archived_code/image_sub.cpp
