+++ src/network_communication/src/net_node.cpp
