+++ src/network_communication/test/assets/test1.png
