+++ src/network_communication/test/full_loop_test.cpp
@@ -1,0 +1,6 @@
+#include <catch2/catchall.hpp>
+#include "client_udp.h"
+#include <vector>
+#include <thread>
+#include <chrono>
+
