--- src/webapp_comms/test/full_loop_test.cpp
@@ -1,6 +1,0 @@
-#include <catch2/catchall.hpp>
-#include "client_udp.h"
-#include <vector>
-#include <thread>
-#include <chrono>
-
